move common functions (IUSaveConfigXXX) to indicom (Common routines)
